main/requirements.txt

alembic     
aniso8601              
bcrypt                 
blinker                
click                  
colorama              
Flask                  
Flask-Bcrypt           
flask-cors             
Flask-JWT-Extended     
flask-marshmallow      
Flask-Migrate          
Flask-RESTful          
Flask-SQLAlchemy       
greenlet               
itsdangerous           
Jinja2                 
Mako                   
MarkupSafe             
marshmallow            
marshmallow-sqlalchemy 
mysql                  
mysqlclient            
packaging              
pip                    
PyJWT                  
python-dotenv         
pytz                 
six                   
SQLAlchemy             
typing_extensions      
Werkzeug               
